% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conExtract.R
\name{conExtract}
\alias{conExtract}
\title{Extract contacts following contact analysis}
\usage{
conExtract(ca, get = "all", idcol = "burst")
}
\arguments{
\item{get}{one of \code{c('all','minDist','minTime')} properties of the contacts to retreive (see Details).}

\item{mtraj}{an object of the class \code{ltraj} which is output from the funciton \code{contacts}.}
}
\value{
A data frame, with the fix attributes associated with contacts.
}
\description{
Create a dataframe with only the attributes associated with contacts, and do some low-level processing.
}
\details{
This function is used to extract contacts from a trajectory following use of the \code{contacts} function. If \code{get='all'} then all contacts are returned. If \code{get='minDist' or 'minTime'}, then where multiple contacts occur for the same fix, only the contact that is nearest in space or time is returned.
}
\seealso{
contacts
}
\keyword{Analysis}
\keyword{Contact}

