% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSegment.R
\name{conSegment}
\alias{conSegment}
\title{Process contact segments}
\usage{
conSegment(ca, sc = 0, idcol = "burst")
}
\arguments{
\item{sc}{time (in seconds) to allow for which to combine contact events (see details).}

\item{traj}{an object of the class \code{ltraj} which is output from the funciton \code{ContactAnalysis}.}
}
\value{
An \code{ltraj} object with new column \code{contact_seg}.
}
\description{
Computes segments where contacts occur based on a temporal tolerance.
}
\details{
This function is used following the \code{ContactAnalysis} function to arrange contacts into segments where continuous contact occurs (based on the user-defined time threshold \code{sc}. The idea is that we can consider a segment to be a continuous contact event (based on \code{dc} see \code{ContactAnalysis}) as long as the contact is only interrupted for no more than \col{sc} time units.
}
\seealso{
dcPlot, contacts, conContext, conSummary
}
\keyword{Analysis}
\keyword{Contact}

