% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSpatial.r
\name{conSpatial}
\alias{conSpatial}
\title{Mapping wildlife contacts}
\usage{
conSpatial(ltraj, type = "p", def = "all",
  proj4string = CRS(as.character(NA)))
}
\arguments{
\item{ltraj}{an object of the class \code{ltraj} which should be output from the function \code{conPhase}.}

\item{type}{one of ('p' - the default or 'l'). Whether to generate contacts as a \code{SpatialPointsDataFrame} or phases as a \code{SpatialLinesDataFrame}, points are the default, but lines can be useful for plotting and exploratory analysis.}

\item{def}{if type = 'p' one of ('all','phase','first','last','minDist','minTime') which defines how contacts are to be mapped using all or part of a contact phase. (see Details)}

\item{proj4string}{a string object containing the projection information to be passed included in the output \code{Spatial*DataFrame} object. For more information see the \code{CRS-class} in the packages \code{sp} and \code{rgdal}. Default is \code{NA}.}
}
\value{
A \code{SpatialPointsDataFrame} or \code{SpatialLinesDataFrame} containing the locations/paths of the contacts. The time of the contact is stored in the attributes of the \code{SpatialPointsDataFrame} object, along with the actual distance between fixes. The \code{SpatialLinesDataFrame} contains attributes of the time of contact, and the min, max, and mean distance apart along a line segment.
}
\description{
The function \code{contacts} is a simple function for mapping where wildlife contacts occur on the landscape with wildilfe telemetry data.
}
\details{
The function \code{conSpatial} can be used to map where contacts occur on the lansdcape, contacts being defined spatially based on a distance threshold \code{dc} and temporally based on the time threshold \code{tc} -- see the function \code{getsimultaneous}. The location of the contact can be calculated in a number of ways, and represented as points for each contact, or as line grouped by the contact phases. Which contacts to map can be defined in a number of ways using the \code{def} parameter: \cr
 \cr i) \code{def = 'all'} (the default) all fixes where column \code{contacts = 1} are returned in the Spatial* object;
 \cr ii) \code{def = 'phase'} all fixes which are part of a phase are returned, note the number of points when \code{def = 'phase'} should be greater than or equal to that when \code{def = 'all'} because of how phases are defined;
 \cr iii) \code{def = 'first'} the first location fix of each phase is returned;
 \cr iv) \code{def = 'last'} the last location fix of each phase is returned;
 \cr v) \code{def = 'minDist'} the location fix of each phase which has the minimal contact distance is returned;
 \cr vi) \code{def = 'minTime'} the location fix of each phase with the minimal time difference with contact fixes is returned;
}
\seealso{
conProcess, conPhase
}
\keyword{Analysis}
\keyword{Contact}
