% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSummary.R
\name{conSummary}
\alias{conSummary}
\title{Summarize contacts and segments}
\usage{
conSummary(cs)
}
\arguments{
\item{cs}{an object of the class \code{ltraj} which should be output from the function \code{conSegment}.}
}
\value{
A dataframe that can be used to summarize contact segments.
}
\description{
Computes some basic summary statistics from a contact analysis.
}
\details{
This function is used following the \code{conSegment} function. It computes the following summary statistics from the contact analysis:
- total number of fixes in the dataset
- total number of fixes deemed a contact
- number of contact segments
- longest segment duration
- mean segment duration
- median segment duration
- no. of segments where the duration is only one fix (i.e., instantaneous contacts)
}
\seealso{
contacts, conSegment
}
\keyword{Analysis}
\keyword{Contact}

