% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimultaneous.R
\name{getSim}
\alias{getSim}
\title{getSim: A faster GetSimultaneous function}
\usage{
getSim(df1, df2, tc = NULL)
}
\arguments{
\item{df1}{a data frame which contains the time-stamped
movement fixes of the first object. This function expects your dataframe to
have columns `date`, `x`, and `y` - in the format of a ltraj data frame.}

\item{df2}{a data frame which contains the time-stamped
movement fixes of the first object.}

\item{tc}{time threshold for determining simultaneous fixes. For simplicity, \code{tc}
is always taken in seconds. If NULL, this adapted function will calculate tolerane on the fly
between 0 and 600 seconds. Intended for use with 15 min, 20 min, and 30 min, regular trajectories.}
}
\value{
a list of two dataframe, each subset to the simultaneous timesteps shared between the dyad.
}
\description{
getSim: A faster GetSimultaneous function
}
