% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj_move2.R
\name{ltraj_move2}
\alias{ltraj_move2}
\title{Convert ltraj to move2 object}
\usage{
ltraj_move2(ltraj)
}
\arguments{
\item{ltraj}{an object of the class \code{ltraj} which contains the time-stamped movement fixes of the object. For more information on objects of this type see \code{help(ltraj)}.}
}
\value{
A \code{move2} object.
}
\description{
The function \code{ltraj_move2} is a simple function for quickly converting \code{ltraj} to \code{move2} objects.
}
\details{
The function \code{ltraj_move2} can be used to convert an \code{ltraj} object into an \code{move2} object.
}
\examples{
data(deer)
deer_ltraj <- move2_ltraj(deer)
deer_move <- ltraj_move2(deer_ltraj)

}
\seealso{
move2_ltraj
}
\keyword{processing}
