% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockhunt.R
\docType{data}
\name{mockhunt}
\alias{mockhunt}
\title{Processed data on contacts between mock-hunters and white-tailed bucks}
\format{An \code{ltraj} object with where bursts represent different individual deer.}
\description{
Data output from the function conContext showing contact events between mock-hunters and male white-tailed deer during the hunting seasons of 2008 and 2009 in Oklahoma.
}
\details{
The mockhunt data are stored as a single \code{dataframe} object; the columns represent environmental attributes associated with contact events.
}
\examples{
data(mockhunt)
head(mockhunt)
}
\references{
Long, JA, Webb, SL, Harju, S, Gee, KL. (in prep) Methods for performing contact analysis from high frequency tracking data: the WildlifeDI R package. Under Review.
}
\keyword{datasets}
