% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move2_ltraj.R
\name{move2_ltraj}
\alias{move2_ltraj}
\title{Convert move2 to ltraj object}
\usage{
move2_ltraj(traj)
}
\arguments{
\item{traj}{an object of the class \code{move2} which contains the time-stamped movement fixes of the object. For more information on objects of this type see \code{help(move2)}.}
}
\value{
A \code{ltraj} object.
}
\description{
The function \code{move2_ltraj} is a simple function for quickly converting \code{move2} to \code{ltraj} objects.
}
\details{
The function \code{move2_ltraj} can be used to convert a \code{move2} object into a \code{ltraj} object.
}
\examples{
data(deer)
deer_ltraj <- move2_ltraj(deer)

}
\seealso{
ltraj_move2
}
\keyword{processing}
