% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does.R
\docType{data}
\name{does}
\alias{does}
\title{GPS tracking data of female white-tailed deer}
\format{An \code{ltraj} object with where bursts represent different individual deer.}
\description{
GPS telemetry data for X does during month of X in 2011.
}
\details{
The doe data are stored as a single \code{'ltraj'} object; each burst represents an individual.
}
\examples{
data(does)
plot(does)
}
\keyword{datasets}
